+++ src/layout/EditTicket.java
@@ -52,6 +52,7 @@ import javax.swing.text.JTextComponent;
 import java.util.Date;
 import java.util.Locale;
 import static layout.Content.areaPanel;
+import static layout.Content.dashboard;
 import static util.Session.*;
 
 /**
@@ -267,26 +268,35 @@ public class EditTicket extends JPanel {
                 } catch (Exception ex) {
                     ex.printStackTrace();
                 }
-                TuyenBUS.editTicket(tuyen.getMaTuyen(), tuyen);
-                areaPanel.removeAll();
-                Dashboard dashBoard = new Dashboard(null);
-                areaPanel.add(dashBoard);
-                areaPanel.validate();
-                areaPanel.repaint();
+                int res = TuyenBUS.editTicket(tuyen.getMaTuyen(), tuyen);
+                if (res == 1) {
+                    JOptionPane.showMessageDialog(null, "Sửa thành công", "Thông báo", 1);
+                    areaPanel.removeAll();
+                    Dashboard dashBoard = new Dashboard(null);
+                    areaPanel.add(dashBoard);
+                    areaPanel.validate();
+                    areaPanel.repaint();
+                } else {
+                    JOptionPane.showMessageDialog(null, "Sửa thất bại", "Thông báo", JOptionPane.ERROR_MESSAGE);
+                }
             }
         });
 
         submitDelete.getButton().addActionListener(new ActionListener() {
             @Override
             public void actionPerformed(ActionEvent e) {
-                TuyenBUS.deleteTicket(tuyen.getMaTuyen());
-                areaPanel.removeAll();
-                Dashboard dashBoard = new Dashboard(null);
-                areaPanel.add(dashBoard);
-                areaPanel.validate();
-                areaPanel.repaint();
-                JOptionPane.showMessageDialog(null, "Xoa thanh cong");
-
+                int check = JOptionPane.showConfirmDialog(null, "Bạn có chắc chắn xóa?", "Thông báo", JOptionPane.OK_CANCEL_OPTION, JOptionPane.WARNING_MESSAGE);
+                if (check == 0) {
+                    int res = TuyenBUS.deleteTicket(tuyen.getMaTuyen());
+                    JOptionPane.showMessageDialog(null, "Xóa thành công", "Thông báo", 1);
+                    if (res != 1) {
+                        JOptionPane.showMessageDialog(null, "Xóa thất bại", "Thông báo", JOptionPane.ERROR_MESSAGE);
+                    }
+                    areaPanel.removeAll();
+                    areaPanel.add(dashboard);
+                    areaPanel.validate();
+                    areaPanel.repaint();
+                }
             }
         });
     }
